WildFly configuration
standalone_bg.sh is a replacement for the WildFly standalone launch script.
It responds to SIGPIPE and thus allows termination of the script once
the process outputs a success message, leaving the server running in a
detached shell.

standalone-ice.xml configures the standalone server to leave the root context
unused and to enable ssl.

jboss-web.xml configures the ice app to run in the root context.

Changes to the makefile in this commit move the files listed above into their
requisite places.