Initial commit of IceApp
This is an incomplete implementation of IceApp, with most of the working
parts in place for the complete implementation.

The original PostgresApp was cloned and stripped of unnecessary bits.
New resources (icons, etc.) were generated for IceApp and all names
migrated in the source. Build scripts were modified to install only
the core of Postgres, and to additionally install WildFly server,
the ICE app, and the necessary BLAST+ binaries.

A class, WildFlyServer, was created to handle management of the WildFly
server process, including starting, stopping, getting status, and managing
apps.

Apart from testing and bug fixes, work still to be done includes:

- add deployment of the ICE war
- add separate reporting of the three statuses (Postgres, WildFly, ICE)
- documentation