Increase app termination timeout to 5 seconds
When shutting down, the app tries to terminate the postgres
and wildfly servers. However, it has a timeout after which
the app will terminate regardless of the state of the subprocesses.

Because we now have both wildly and postgres, where before there was
only postgres, it makes sense to wait a bit longer. Here we increase
the timeout to 5 seconds.