Move to direct postgres compile
Previously we relied on the postgres that comes packaged with
Postgres.app. However, this includes hard-coded references to
the /Applications/Postgres.app directory.

To solve this we have moved to direct compilation of postgres
during make, with all hard-coded references now pointing to
/Applications/Ice.app. To accomodate this, the makefile from
Postgres.app was modified to only compile the core postgres
binaries we want, and all references to Postgres.app in the
objective-c source were converted to Ice.app.