Fix postgres compilation
Postgres compilation was broken in several ways:

1. the makefile tries to build docs which require unnecessary dependencies
2. there were problems with the $libdir variable in compilation (see #12)

Both problems are solved by issuing sed commands from the makefile
to replace the troublesome parts of files.

In the case of 1, all dependencies on doc within the pre-configuration makefile
are removed.

In the case of 2, all mentions of $libdir are replaced with the full path to the
installed lib directory, within the Ice.app.