Try to create user DB even if data dir exists
In Postgres.app, the user DB is only created if the data dir
does not exist. However, this causes issues when testing during
development. This commit ensures the user DB creation will be
attempted even if the data dir exists, allowing more accurate
development running of the app. It should not affect deployment
running.